+++ BackEnd/com/src/main/kotlin/example/com/models/Task.kt
@@ -1,12 +1,13 @@
 package example.com.models
 
-import org.jetbrains.exposed.dao.IntIdTable
+import org.jetbrains.exposed.dao.id.IntIdTable
+import org.jetbrains.exposed.sql.ReferenceOption
 import org.jetbrains.exposed.sql.Table
 
 object TaskTable : IntIdTable() {
     val title = varchar("title", 255)
     val description = text("description")
-    val userId = reference("user_id", UserTable)
+    val userId = reference("user_id", UserTable,  onDelete = ReferenceOption.CASCADE)
 }
 
 data class Task (