+++ BackEnd/com/src/main/kotlin/example/com/models/User.kt
@@ -1,6 +1,6 @@
 package example.com.models
 
-import org.jetbrains.exposed.dao.IntIdTable
+import org.jetbrains.exposed.dao.id.IntIdTable
 import org.jetbrains.exposed.sql.Table
 
 object UserTable : IntIdTable() {